Changes the input value for the dispatchAfter helper function from Int64 to Double.

+++ GreatStuff/Structs/GCD/GSGCD.swift
@@ -15,8 +15,8 @@ public struct GSGCD {
      
          - Note: in case _queue_ is not provided, then _dispatch_get_main_queue()_ is used by default.
      */
-    public static func dispatchAfter(sec: UInt64, queue: dispatch_queue_t = dispatch_get_main_queue(), block: () -> Void) {
-        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, Int64(sec * NSEC_PER_SEC)), queue, block)
+    public static func dispatchAfter(sec: Double, queue: dispatch_queue_t = dispatch_get_main_queue(), block: () -> Void) {
+        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, Int64(sec * Double(NSEC_PER_SEC))), queue, block)
     }
     
 }
